+++ test/redox_test.rb
@@ -15,6 +15,11 @@ class RedoxTest < Minitest::Test
       .with(body: hash_including(request_body),
             headers: { 'Authorization' => 'Bearer let.me.in' })
       .to_return(status: 200, body: '{ "Success": true }')
+
+    stub_request(:post, 'https://api.redoxengine.com/endpoint')
+      .with(body: hash_including(request_body('NewPatient')),
+            headers: { 'Authorization' => 'Bearer let.me.in' })
+      .to_return(status: 200, body: '{ "Success": true }')
   end
 
   def test_add_patient
@@ -29,13 +34,25 @@ class RedoxTest < Minitest::Test
     redox.add_patient(patient)
   end
 
+  def test_update_patient
+    redox = Redox::Redox.new(
+      api_key: '1-2-3',
+      secret: 'xJp3Ba',
+      source: source,
+      destinations: destinations,
+      test: true
+    )
+
+    redox.update_patient(patient)
+  end
+
   private
 
-  def request_body
+  def request_body(event_type='PatientUpdate')
     {
       Meta: hash_including(
         DataModel: 'PatientAdmin',
-        EventType: 'PatientUpdate',
+        EventType: event_type,
         Test: true,
         Source: source,
         Destinations: destinations